Move Host related entity classes under com.futurewei.alcor.web.entity.node
