+++ services/node_manager/src/main/java/com/futurewei/alcor/nodemanager/controller/NodeController.java
@@ -25,8 +25,8 @@ import com.futurewei.alcor.common.utils.RestPreconditionsUtil;
 import com.futurewei.alcor.nodemanager.exception.InvalidDataException;
 import com.futurewei.alcor.nodemanager.service.NodeService;
 import com.futurewei.alcor.nodemanager.utils.NodeManagerConstant;
-import com.futurewei.alcor.web.entity.NodeInfo;
-import com.futurewei.alcor.web.entity.NodeInfoJson;
+import com.futurewei.alcor.web.entity.node.NodeInfo;
+import com.futurewei.alcor.web.entity.node.NodeInfoJson;
 import com.futurewei.alcor.web.entity.node.BulkNodeInfoJson;
 import com.futurewei.alcor.web.entity.node.NodesWebJson;
 import com.futurewei.alcor.web.json.annotation.FieldFilter;
