+++ services/port_manager/src/main/java/com/futurewei/alcor/portmanager/entity/PortBindingHost.java
@@ -15,7 +15,7 @@ Licensed under the Apache License, Version 2.0 (the "License");
 */
 package com.futurewei.alcor.portmanager.entity;
 
-import com.futurewei.alcor.web.entity.NodeInfo;
+import com.futurewei.alcor.web.entity.node.NodeInfo;
 
 public class PortBindingHost {
     private String portId;
