+++ web/src/main/java/com/futurewei/alcor/web/entity/node/NodeInfo.java
@@ -9,7 +9,7 @@ Licensed under the Apache License, Version 2.0 (the "License");
         See the License for the specific language governing permissions and
         limitations under the License.
 */
-package com.futurewei.alcor.web.entity;
+package com.futurewei.alcor.web.entity.node;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonProperty;
