+++ web/src/main/java/com/futurewei/alcor/web/entity/node/NodeInfoJson.java
@@ -9,9 +9,10 @@ Licensed under the Apache License, Version 2.0 (the "License");
         See the License for the specific language governing permissions and
         limitations under the License.
 */
-package com.futurewei.alcor.web.entity;
+package com.futurewei.alcor.web.entity.node;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.futurewei.alcor.web.entity.node.NodeInfo;
 import lombok.Data;
 
 @Data
