+++ web/src/main/java/com/futurewei/alcor/web/entity/node/NodesWebJson.java
@@ -15,8 +15,6 @@ Licensed under the Apache License, Version 2.0 (the "License");
 */
 package com.futurewei.alcor.web.entity.node;
 
-import com.futurewei.alcor.web.entity.NodeInfo;
-
 import java.util.List;
 
 public class NodesWebJson {
